+++ pydash/functions.py
@@ -1,0 +1,2 @@
+"""Functions
+"""
