+++ pydash/objects.py
@@ -1,0 +1,2 @@
+"""Objects
+"""
